Chatbot connected to server and Fin.
Typically the react app can be started by running 'npm start', this
should start up the app on http://localhost:3000/.
The python flask server runs on http://localhost:5000/, and that has
been used in axios post request as well. This IP address can be changed.

To confirm server response is coming to chatbot, strings are reversed
in flask server when getting sent.